+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Feat
+
+- opencv and unstructured
+
 ## v0.1.1 (2023-02-22)
 
 ## v0.1.0 (2023-02-22)